// ui/src/generated/api/model/dtoFileResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoFileResponseKind } from "./dtoFileResponseKind";

export interface DtoFileResponse {
  expectedVerdict: string;
  id: number;
  isActive: boolean;
  kind: DtoFileResponseKind;
  language: string;
  name: string;
  sourceCode: string;
  updatedAt: string;
}