// ui/src/generated/api/model/dtoStatementResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export interface DtoStatementResponse {
  inputFormat: string;
  language: string;
  legend: string;
  name: string;
  notes: string;
  outputFormat: string;
  scoring: string;
  tutorial: string;
  updatedAt: string;
}