// ui/src/generated/api/model/dtoTestResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoTestResponseSource } from "./dtoTestResponseSource";

export interface DtoTestResponse {
  description: string;
  generateCmd: string;
  group: string;
  id: number;
  index: number;
  inputData: string;
  isSample: boolean;
  points: number;
  source: DtoTestResponseSource;
}